feat: rename Java class
